Move setup_platform to __init__.py and patch HA object instead of library
