+++ tests/components/openweathermap/test_weather.py
@@ -10,7 +10,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ServiceValidationError
 from homeassistant.helpers import entity_registry as er
 
-from .conftest import setup_platform
+from . import setup_platform
 
 from tests.common import MockConfigEntry, snapshot_platform
